requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pymongo
python-jose
passlib[bcrypt]
